+++ src/Configuration/ObjectConfiguration.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Sobak\Scrawler\Configuration;
 
-use ArrayIterator;
 use Sobak\Scrawler\Block\ResultWriter\ResultWriterInterface;
 use Sobak\Scrawler\Matcher\ListMatcherInterface;
 use Sobak\Scrawler\Matcher\MatcherInterface;
@@ -18,7 +17,7 @@ class ObjectConfiguration
 
     protected $matcher;
 
-    /** @var ResultWriterInterface[] */
+    /** @var array */
     protected $resultWriters = [];
 
     public function __construct(ListMatcherInterface $matcher)
@@ -71,7 +70,7 @@ class ObjectConfiguration
 
     public function addResultWriter(string $entityClass, ResultWriterInterface $resultWriter): self
     {
-        $this->resultWriters[$entityClass] = $resultWriter;
+        $this->resultWriters[$entityClass][] = $resultWriter;
 
         return $this;
     }
@@ -79,12 +78,5 @@ class ObjectConfiguration
     public function getResultWriters()
     {
         return $this->resultWriters;
-    }
-
-    public function removeResultWriter(string $entityClass): self
-    {
-        unset($this->resultWriters[$entityClass]);
-
-        return $this;
     }
 }